build 390 tested from repo BCC, VS2008 all, VS2015
git-svn-id: svn://svn.code.sf.net/p/steemsse/code/trunk@550 dd3fbd3b-ef5d-4b78-98e8-da144fa02e5b

+++ steemsse/3rdparty/avi/AviFile.cpp
@@ -16,7 +16,7 @@
 #define _tcsncpy strncpy
 extern IDirectSoundBuffer *PrimaryBuf,*SoundBuf; // VC9 doesn't know that
 unsigned int FormatAviMessage(HRESULT code, char *buf,unsigned int len);
-extern int shifter_freq_at_start_of_vbl;
+extern BYTE shifter_freq_at_start_of_vbl;
 
 #include "avifile.h"
